deepen: AnchorPipeline owns autofix chain (inline parse/resolve/apply)
Inline parse (parseHashRef/parseText, 50), resolve (resEdit/stripBare/
stripDiff/swapReversed/valEdit/boundaryDups, 544) and apply (applyEdit/
resToSpan/assemble, 349) and served (verifyServedRange, 194) into
anchor-pipeline.ts (now 1121 lines, deep). Single ordering invariant
swapReversed→stripBare→stripDiff→valEdit→boundaryDups splice→valEdit→
verifyServed→resToSpan now co-locates. Detection (valEdit→boundaryDups)
and correction (splice+second valEdit) no longer split with implicit
second-valEdit coupling — one test surface.

Shallow files parse/resolve/apply/served become thin shims re-exporting
from anchor-pipeline for compat; stripBarePrefixes etc. are private to
the seam (not exported from anchor-pipeline, except findNewEdge kept
@internal for fuzz test). Typecheck + 674 tests pass.
