+++ src/hashline/parse.ts
@@ -1,50 +1,2 @@
-import {
-ANCHOR_LEN,
-	ALPH_RE,
-} from "./pure.js";
-import { NEW_CONTENT_NOT_STRING_MSG } from "../constants.js";
-
-export type Anchor = { hash: string };
-
-function diagRef(ref: string): string {
-	const trimmed = ref.trim();
-
-	if (!trimmed.length) {
-		return `[E_BAD_REF] Invalid anchor. Expected a 3-char alphanumeric anchor (e.g. "aB3").`;
-	}
-
-	if (/^\d+/.test(trimmed)) {
-		return `[E_BAD_REF] Invalid anchor. Use the hash alone (e.g. "aB3") — no line numbers or trailing content.`;
-	}
-
-	if (trimmed.includes("│")) {
-		return `[E_BAD_REF] Invalid anchor "${trimmed}". remove_from and remove_to must contain the 3-char hash only — remove everything from "│" onward.`;
-	}
-
-	return `[E_BAD_REF] Invalid anchor "${trimmed}". Expected a 3-char alphanumeric anchor (e.g. "aB3").`;
-}
-
-function parseRef(ref: string): Anchor {
-	const trimmed = ref.trim();
-
-	if (
-		trimmed.length === ANCHOR_LEN &&
-		ALPH_RE.test(trimmed)
-	) {
-		return { hash: trimmed };
-	}
-
-	throw new Error(diagRef(ref));
-}
-
-export const parseHashRef = parseRef;
-
-export function parseText(edit: string): string[] {
-  if (typeof edit !== "string") {
-    throw new Error(NEW_CONTENT_NOT_STRING_MSG);
-  }
-  const normalized = edit.replace(/\r\n/g, "\n").replace(/\r/g, "\n");
-  if (normalized === "") return [];
-  if (/^\n+$/.test(normalized)) return new Array(normalized.length).fill("");
-  return normalized.split("\n");
-}
+// @deprecated — private to AnchorPipeline seam. Use `from "./anchor-pipeline.js"` instead.
+export * from "./anchor-pipeline.js";
